// server/src/shared/types/db.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Role = "ADMIN" | "USER";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Activity {
  action: string;
  createdAt: Generated<Timestamp>;
  id: string;
  metadata: Json | null;
  type: string;
  userId: string;
}

export interface Session {
  createdAt: Generated<Timestamp>;
  deletedAt: Timestamp | null;
  expiresAt: Timestamp;
  id: string;
  lastLogin: Timestamp | null;
  refreshToken: string;
  updatedAt: Timestamp;
  userId: string;
}

export interface User {
  avatar: Generated<string | null>;
  createdAt: Generated<Timestamp>;
  deletedAt: Timestamp | null;
  email: string;
  firstName: string;
  id: string;
  lastName: string;
  password: string;
  role: Generated<Role>;
  updatedAt: Timestamp;
}

export interface DB {
  Activity: Activity;
  Session: Session;
  User: User;
}